+++ acro/packages/colin/src/colin/reformulation/UnconMultiobj.h
@@ -347,8 +347,8 @@ private: // methods
       while ( row > 0 )
       {
          --row;
-         size_t i = cg.get_matbeg()[row];
-         size_t n = cg.get_matcnt()[row];
+         size_t i = static_cast<size_t>(cg.get_matbeg()[row]);
+         size_t n = static_cast<size_t>(cg.get_matcnt()[row]);
          while ( n > 0 )
          {
             g[cg.get_matind()[i]] += 2.0*cv[row]*cg.get_matval()[i];
